Removed dependency on mintinstall

+++ usr/lib/linuxmint/mintMenu/plugins/system_management.py
@@ -97,11 +97,12 @@ class pluginclass( object ):
 	#Add standard items
 	def do_standard_items( self ):		
 
-		Button1 = easyButton( "/usr/lib/linuxmint/mintInstall/icon.svg", self.iconsize, [_("Software manager")], -1, -1 )
-		Button1.connect( "clicked", self.ButtonClicked, "mintinstall" )
-		Button1.show()
-		self.systemBtnHolder.pack_start( Button1, False, False )
-		self.mintMenuWin.setTooltip( Button1, _("Browse and install available software") )
+		if os.path.exists("/usr/lib/linuxmint/mintInstall/icon.svg"):
+			Button1 = easyButton( "/usr/lib/linuxmint/mintInstall/icon.svg", self.iconsize, [_("Software manager")], -1, -1 )
+			Button1.connect( "clicked", self.ButtonClicked, "mintinstall" )
+			Button1.show()
+			self.systemBtnHolder.pack_start( Button1, False, False )
+			self.mintMenuWin.setTooltip( Button1, _("Browse and install available software") )
 
 		Button2 = easyButton( "synaptic", self.iconsize, [_("Package manager")], -1, -1 )
 		Button2.connect( "clicked", self.ButtonClicked, "gksu /usr/sbin/synaptic" )
